qlora

+++ opendelta/delta_models/lora.py
@@ -32,8 +32,8 @@ class LowRankLinear(nn.Module):
             self.lora_dropout = lambda x: x
         if r > 0:
             ### change here
-            self.lora_A = nn.Parameter(weight.half().new_zeros((r, in_features)))
-            self.lora_B = nn.Parameter(weight.half().new_zeros((out_features, r)))
+            self.lora_A = nn.Parameter(weight.float().new_zeros((r, in_features)))
+            self.lora_B = nn.Parameter(weight.float().new_zeros((out_features, r)))
             self.scaling = self.lora_alpha / self.r
             nn.init.kaiming_uniform_(self.lora_A, a=math.sqrt(5))
             nn.init.zeros_(self.lora_B)
